Some doc updating, more schema migration stuff
