+++ lib/aqueduct.dart
@@ -29,10 +29,14 @@ import 'package:crypto/crypto.dart';
 import 'package:logging/logging.dart';
 import 'package:matcher/matcher.dart';
 import 'package:postgres/postgres.dart';
+import 'package:args/args.dart';
+import 'package:safe_config/safe_config.dart';
 
 export 'package:args/args.dart';
 export 'package:logging/logging.dart';
 export 'package:safe_config/safe_config.dart';
+export 'dart:io';
+export 'dart:async';
 
 part 'auth/auth_code_controller.dart';
 part 'auth/auth_controller.dart';
@@ -63,6 +67,9 @@ part 'base/response.dart';
 part 'base/route_node.dart';
 part 'base/router.dart';
 part 'base/serializable.dart';
+part 'commands/template_creator.dart';
+part 'commands/cli_command.dart';
+part 'commands/migration_runner.dart';
 part 'db/data_model.dart';
 part 'db/data_model_builder.dart';
 part 'db/matcher_expression.dart';
@@ -82,6 +89,8 @@ part 'db/predicate.dart';
 part 'db/query.dart';
 part 'db/query_page.dart';
 part 'db/schema.dart';
+part 'db/schema_table.dart';
+part 'db/schema_column.dart';
 part 'db/schema_builder.dart';
 part 'db/sort_descriptor.dart';
 part 'utilities/mirror_helpers.dart';
