+++ pubspec.yaml
@@ -16,8 +16,11 @@ dependencies:
   logging: ">=0.11.3 <0.12.0"
   analyzer: ">=0.28.0 <0.30.0"
   postgres: ">=0.9.0 <1.0.0"
-  args: ">=0.13.5 < 1.0.0"
+  args: ">=0.13.5 <1.0.0"
 
 dev_dependencies:
   test: any
   http: ">=0.11.3+7 <0.12.0"
+
+executables:
+  aqueduct: aqueduct
