+++ test/db/schema_test.dart
@@ -1,0 +1,6 @@
+import 'package:aqueduct/aqueduct.dart';
+import 'package:test/test.dart';
+
+void main() {
+
+}
